VS codium
This function processes an array of student objects. It calculates the average score for each student and then filters out the students whose average score is greater than 85. Let's go through it:

1. Define an array of student objects :
2. Calculate average scores: 
3. Filter students with scores above 85:
4. Logging:

When you run this function, it will output the original array of student data and then the list of students with average scores above 85.